% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-caller.R
\name{export_ptt}
\alias{export_ptt}
\title{Export data store in miner_object for you to analyze.}
\usage{
export_ptt(ptt.miner, export.type = character(), obj.name = character())
}
\arguments{
\item{ptt.miner}{a R6 class object uses \code{PTTmineR$new()} to create.}

\item{export.type}{a string. the three export types in description part.}

\item{obj.name}{a string. the global object name you want to name.}
}
\value{
\code{export_ptt()} is a side-effect function also. After you excute
this function, you will get the object name 'obj.name' in your
global environment.
}
\description{
There're three export types you can choose:
\itemize{
\item dt: the type of data.table
\item tbl: the type of tibble (like data.frame but more better)
\item nested_tbl: the type of nested tibble (just return one table)
}
}
\examples{
# export data into global environment.
\dontrun{rookie_miner \%>\%
    export_ptt(export.type = "dt", obj.name = "my_ptt_data")
    }

# then you can call the object you name in the global environment.
\dontrun{my_ptt_data}

}
