% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-caller.R
\name{mine_ptt}
\alias{mine_ptt}
\title{Crawl the selected board post and store into PTTmineR.}
\usage{
mine_ptt(
  ptt.miner,
  board = NULL,
  keyword = NULL,
  author = NULL,
  recommend = NULL,
  min.date = NULL,
  last.n.page = NULL
)
}
\arguments{
\item{ptt.miner}{a R6 class object uses \code{PTTmineR$new()} to create.}

\item{board}{a string. PTT board you want to crawl.}

\item{keyword}{a string. The keyword you want to search on selected board.}

\item{author}{a string. The post author you want to search on selected board.}

\item{recommend}{a number. The number of net recommend you want to filter.}

\item{min.date}{a date format string. The farthest date you want to set.}

\item{last.n.page}{a number. The number of page you wnat to crawl.}
}
\value{
\itemize{
\item post_info_dt: the data.table type post's information.
\item post_comment_dt: the data.table type post's comments.
The result is store in miner_object$result_dt, but my
suggestion is don't modify the data in miner_object, use
\code{export_ptt()} to get and analysis your data.
}
}
\description{
You can use following arguments to get/filter the
PTT post data you want. Note that you need at least
choose which board you want to crawl.
}
\examples{

# assume that rookie_miner is an object using `PTTmineR$new()` to create
# get all Gossiping posts
\dontrun{rookie_miner \%>\%
    mine_ptt(board = "Gossiping")}

# get all Gossiping posts to filter by keyword 'youtuber'
\dontrun{rookie_miner \%>\%
    mine_ptt(board = "Gossiping", keyword = "youtuber")}

# get all Gossiping posts to filter by keyword 'youtuber' and
# net recommend nuber 10.
\dontrun{rookie_miner \%>\%
    mine_ptt(board = "Gossiping", recommend = 10)}

# if you want to do multiple crawling task on one eval,
# you can use multiple `\%>\%` :
\dontrun{rookie_miner \%>\%
    mine_ptt(board = "Gossiping", recommend = 10) \%>\%
    mine_ptt(board = "Soft_job", keyword = "python")}

# or use `purrr::pwalk()`:
\dontrun{board_list <- c("Gossiping", "Soft_job", "Beauty")
pwalk(board_list, ~mine_ptt(board = .x, recommend = 10))}
# why `pwalk()`? because all PTTmineR's functions are
# side-effect funciton. The data will return to rookie_miner.



}
