% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-option.R
\name{plan}
\alias{plan}
\alias{multiprocess}
\alias{sequential}
\title{future option}
\description{
See \code{future::\link[future:plan]{plan}} for details.

See \code{future::\link[future:multiprocess]{multiprocess}} for details.

See \code{future::\link[future:sequential]{sequential}} for details.
}
\keyword{internal}
