% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTTmineR.R
\name{PTTmineR}
\alias{PTTmineR}
\title{Create a PTTmineR object}
\description{
Imaging that PTTmineR object is a data center. You can store the data
on each crawling task. This design method is to make it easier for
users to manage the data they crawl on each crawling task.
}
\examples{
# create PTTmineR object
\dontrun{rookie_miner <- PTTmineR$new(task.name = "first_task")}

# print the summary information of rookie_miner
\dontrun{rookie_miner}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{result_dt}}{a list of following result data.table:
\itemize{
\item post_info_dt: the data.table type post's information.
\item post_comment_dt: the data.table type post's comments.
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PTTmineR$new()}}
\item \href{#method-print}{\code{PTTmineR$print()}}
\item \href{#method-clone}{\code{PTTmineR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
overrides the default behaviour of \verb{$new()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTTmineR$new(task.name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task.name}}{a string. the name of miner's task}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
print the summary information of PTTmineR object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTTmineR$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTTmineR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
