% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-caller.R
\name{update_ptt}
\alias{update_ptt}
\title{Update the post comments have been crawled.}
\usage{
update_ptt(ptt.miner, update.post.id = character())
}
\arguments{
\item{ptt.miner}{a R6 class object uses `PTTmineR$new()` to create.}

\item{update.post.id}{the charactar vector post id you want to update.}
}
\description{
Update the post comments have been crawled.
}
\examples{
# get the post id set from rookie_miner
\dontrun{update_id <- rookie_miner$result_dt$post_info_dt$post_id}

# update the post comments from selected post id
\dontrun{rookie_miner \%>\%
    update_ptt(update.post.id = update_id)}
}
